--- src/web/common/constants.js
@@ -1,14 +1,0 @@
-/* src/web/aws-cloudwatch/reducers/StepsContext.jsx */
-const ADD_ENABLED_STEP = 'ADD_ENABLED_STEP';
-const SET_AVAILABLE_STEPS = 'SET_AVAILABLE_STEPS';
-const SET_CURRENT_STEP = 'SET_CURRENT_STEP';
-
-/* src/web/aws-cloudwatch/reducers/FormDataContext.jsx */
-// const UPDATE_FORM_DATA = 'UPDATE_FORM_DATA';
-
-export {
-  ADD_ENABLED_STEP,
-  SET_AVAILABLE_STEPS,
-  SET_CURRENT_STEP,
-  // UPDATE_FORM_DATA,
-};
